Modify: 일부 리펙토링
